add pagination (user & group) and correct routes

+++ apps/api/src/usecases/auth/accountManagement.usecases.ts
@@ -1,13 +1,12 @@
-import { BadRequestException } from "@nestjs/common";
+import { BadRequestException, NotFoundException } from "@nestjs/common";
 import { ILogger } from "../../domain/logger/logger.interface";
 import { IAccountRepository } from "../../domain/repositories/accountRepository.interface";
 import Role from "../../domain/models/enums/role.enum";
 import { IRolePermissionsRepository } from "../../domain/repositories/rolePermissionsRepository.interface";
 import Permission from "../../domain/models/enums/permission.type";
 import { AccountMapper } from "../../infrastructure/mappers/account.mapper";
-import { NestedAccountM } from "../../domain/models/account";
-import { Account } from "../../infrastructure/entities/Account.entity";
 import { PaginateQuery, Paginated } from "nestjs-paginate";
+import { UserAccountSummaryDto } from "../../infrastructure/controllers/userManagement/userAccountDto.class";
 
 export class AccountManagementUseCases {
     constructor(
@@ -23,7 +22,7 @@ export class AccountManagementUseCases {
         const account = await this.accountRepository.getAccountByUsername(username);
         if (!account) {
             this.logger.error('AccountManagementUseCases accountIsValid', 'Account not found')
-            throw new BadRequestException('Account not found');
+            throw new NotFoundException('Account not found');
         }
 
         return account.active;
@@ -36,7 +35,7 @@ export class AccountManagementUseCases {
         const account = await this.accountRepository.getAccountByUsername(username);
         if (!account) {
             this.logger.error('AccountManagementUseCases updateAccountState', 'Account not found')
-            throw new BadRequestException('Account not found');
+            throw new NotFoundException('Account not found');
         }
 
         await this.accountRepository.updateAccountState(username, !account.active);
@@ -72,8 +71,13 @@ export class AccountManagementUseCases {
     /// </summary>
     async deleteAccount(id: string): Promise<any> {
         const account=await this.accountRepository.findAccountById(id);
+
+        if(!account){
+            throw new NotFoundException("account not found");
+        }
+
         if(account.active){
-            throw new BadRequestException("account is active");
+            throw new BadRequestException("account is active and cannot be deleted");
         }
         const role=account.rolePermissions.role;
         return await this.accountRepository.deleteAccount(id);
@@ -82,20 +86,19 @@ export class AccountManagementUseCases {
     /// <summary>
     ///     Find an account
     /// </summary>
-    async findAccountDetails(id: string, username: string, firstname: string, lastname: string): Promise<any> {
-        let account = await this.accountRepository.findAccountDetails(id, username, firstname, lastname);
-        if (!account) {
+    async findAccountDetails(accountId: string): Promise<any> {
+        let userAccountDetails = await this.accountRepository.findUserAccountDetails(accountId);
+        if (!userAccountDetails) {
             this.logger.error('AccountManagementUseCases findAccountDetails', 'Account not found')
-            throw new BadRequestException('Account not found');
+            throw new NotFoundException('Account not found');
         }
-
-        return account;
+        return userAccountDetails;
     }
 
     /// <summary>
     ///    Find accounts with pagination
     /// </summary>
-    async findAll(query: PaginateQuery): Promise<Paginated<Account>> {
-        return await this.accountRepository.findAll(query);
+    async findAll(query: PaginateQuery): Promise<any> {
+        return await this.accountRepository.findAccount(query);
     }
 }
